pyproject.toml
[tool.poetry]
name = "goit-py-web-hw-06"
version = "0.1.0"
description = ""
authors = ["Serge-Moskalenko <skmos2019@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
faker = "^33.1.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
